// frontend/src/api/generated/api.ts
export interface Account {
    'accountId'?: number;
    'customerId'?: number;
    'productCode'?: string;
    'accountNumber'?: string;
    'balance'?: number;
    'status'?: string;
}
export interface AccountsIdPutRequest {
    'status'?: string;
}
export interface Customer {
    'customerId'?: number;
    'firstName'?: string;
    'lastName'?: string;
    'email'?: string;
    'cifNumber'?: string;
    'dateOfBirth'?: string;
    'kycStatus'?: string;
    'riskRating'?: string;
    'accountCount'?: number;
    'accounts'?: Array<Account>;
}
export interface Transaction {
    'transactionId'?: number;
    'accountId'?: number;
    'transactionType'?: string;
    'amount'?: number;
    'transactionDate'?: string;
    'balanceAfter'?: number;
}